refactor prints into new file
